Completed the library management system

+++ Module-3/Library Management System/add_book.py
@@ -4,7 +4,7 @@ def add_books(all_books):
     title = input("Enter the Book Title: ")
     author = input("Enter the Author Name: ")
     isbn = int(input("Enter the ISBN: "))
-    price = int(input("Enter the Book Price: "))
+    price = float(input("Enter the Book Price: "))
     year = int(input("Enter the publishing Year: "))
     quantity = int(input("Enter the quantity: "))
 
@@ -19,7 +19,6 @@ def add_books(all_books):
     
     all_books.append(book)
     save_all_books(all_books)
-
     print("Book added successfully.")
 
     return all_books
